requirements.txt
streamlit>=1.30.0
pandas
google-generativeai>=0.8.0
google-auth
google-auth-oauthlib
google-api-python-client
st-gsheets-connection
pypdf
Pillow